Add files via upload
The HPE project aims to revolutionize driving safety  by developing a system that 
accurately identifies and interprets drivers' behaviour and posture using real-time computer vision 
techniques. Recognizing the increasing concerns over road safety due to driver distractions, our 
system employs advanced pose estimation methods to monitor priority joints and postures of 
drivers, highlighting potential distractions or drowsiness. By integrating this technology into 
vehicles, we aim to proactively enhance driving safety, reduce accident rates, and prevent 
fatalities. 
We pioneered the use of the YOLOv8 model for its accuracy and speed for driver’s safety, making 
it ideal for estimating driver posture. Utilizing an open-source dataset of drivers initially encoded 
in YAML format, we formatted it to be YOLO-compatible, which was perfect for training 
YOLOv8 for our pose estimation project. Our model was rigorously tested for dynamic stability 
on numerous still images and real-life scenarios, enabling our system to effectively capture and 
address driver behaviours, thereby enhancing the driver monitoring and safety alert features. 
The HPE project represents a significant technological advancement and contributes to improving 
road safety and overall societal welfare. As the first to use and train the DriPE dataset on the 
YOLOv8 model, our project demonstrates the potential of intelligent systems to make a positive 
impact. 
We continue to refine and expand our project based on testing feedback and ongoing 
advancements in computer vision and machine learning. Our goal is to mainstream this 
technology, making roads safer not only in Pakistan but globally. We hope this project exemplifies 
the transformative power of technology when harnessed to improve human lives.